product 1.1 login 1.1

+++ src/com/grandet/controller/CollectionController.java
@@ -55,11 +55,11 @@ public class CollectionController {
     public @ResponseBody
     Map<String, String> addCollection(Collection collection, HttpServletRequest request){
         Map<String, String > map = new HashMap<String, String>();
-//        User user = (User)request.getSession().getAttribute("currentUser");
-//        if (user.getId()!=collection.getUserId()){
-//            map.put("msg", "not match");
-//            return map;
-//        }
+        User user = (User)request.getSession().getAttribute("currentUser");
+        if (user.getId()!=collection.getUserId()){
+            map.put("msg", "not match");
+            return map;
+        }
         int result = collectionService.addCollection(collection);
         if (result == 1){
             map.put("msg", "success");
@@ -85,11 +85,11 @@ public class CollectionController {
             map.put("msg", "not exist");
             return map;
         }
-//        User user = (User)request.getSession().getAttribute("currentUser");
-//        if (user.getId()!=collection.getUserId()){
-//            map.put("msg", "not match");
-//            return map;
-//        }
+        User user = (User)request.getSession().getAttribute("currentUser");
+        if (user.getId()!=collection.getUserId()){
+            map.put("msg", "not match");
+            return map;
+        }
         int result = collectionService.deleteCollection(id);
         if (result == 1){
             map.put("msg", "success");
